Bug Fix: When the categoryField is not specified on the Category?Search component, auto suggest is broken. Given that the category search field is optional according to the documentation here https://opensource.appbase.io/reactive-manual/search-components/categorysearch.html#props
This should have worked, but the code that recommends the final suggesiton list is only merging the two results when category search is enabled. This can be replicated in the demo example from the documentation site by removing the category field. I have created a codesandbox.io that showcases the broken search functionality here
https://codesandbox.io/embed/2orqz6rryn

The fix ensures the the parsedSuggestions getter always merges the categoriesResults and the suggestionsResults.