Remove error log when URL param cannot be decoded
For scenarios where reactivesearch is part of a larger application that uses URL parameters that are not in JSON format we do not want to log errors since it is expected that some URL parameters will not belong to reactivesearch. See #955 for associated Vue issue that was resolved in a similar manner.