fix(vue): logic to build default SingleList filter query
Fixes https://github.com/appbaseio/reactivesearch/issues/1481
The query that handles the case when the user selects to find items with no value for the filter field gets overwritten by the query that uses the selected term value.

The solution is to create the term query by default, and overwrite it with the missing value query if necessary.