fix(vue): `ReactiveList.updateQueryOptions` ignores `defaultSortOption`
ReactiveList executes method `updateQueryOptions()` when some properties get modified. However, it ignores the `defaultSortOption` property when it is set.

In [this code sandbox](https://codesandbox.io/s/wispy-night-o1rj6) we added properties `sortOptions` and `defaultSortOption`.

1. Notice, that when loading the page:
    * it shows the Sort selector, which has 3 options:
        * Title (descending)
        * Relevance (descending)
        * Rating (average rating descending)
    * it shows Rating as the default sort order
        * as specified by property `defaultSortOption` passed to component ReactiveList.
    * items are sorted by average rating accordingly
1. Now when we submit a search term:
    * this code modifies property `dataField` of the `ReactiveList` component, via `data.sort`
    * in turn, this triggers the execution of method `ReactiveList.updateQueryOptions()`

*Note:* the search is executed however:
* the results are sorted by Title (descending)
* whereas the selected Sort option is still the default (Rating)

This is because `ReactiveList.updateQueryOptions()` is always taking the first sort options, which is not necessarily the default one.

We find and use the default Sort Option to set the sort Query option.